requirements.txt
streamlit==1.41.1
spotipy==2.24.0
Pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1